+++ chartlets.js/src/lib/actions/handleHostStoreChange.ts
@@ -79,7 +79,7 @@ function getHostStorePropertyRefs(): PropertyRef[] {
       (contribution.callbacks || []).forEach(
         (callback, callbackIndex) =>
           (callback.inputs || []).forEach((input, inputIndex) => {
-            if (!input.noTrigger && input.link === "app" && input.property) {
+            if (!input.noTrigger && input.id === "@app" && input.property) {
               propertyRefs.push({
                 contribPoint,
                 contribIndex,