+++ chartlets.js/src/lib/types/model/channel.ts
@@ -1,21 +1,17 @@
 import type { ObjPathLike } from "@/lib/utils/objPath";
 
-export type Link = "component" | "container" | "app";
-
 /**
  * Base for `Input` and `Output`.
  */
 export interface Channel {
   /**
-   * The link provides the source for inputs and that target for outputs.
+   * The identifier for a component or state.
+   *
+   * Special identifiers are:
+   * - `"@app"` the application state referred to by `HostStore`
+   * - `"@container"` the state referred to by contribution's container
    */
-  link: Link;
-
-  /**
-   * The identifier of a subcomponent.
-   * `id` is not needed if link == "AppInput" | "AppOutput".
-   */
-  id?: string;
+  id: "@app" | "@container" | string;
 
   /**
    * The property of an object or array index.
@@ -28,3 +24,15 @@ export interface Input extends Channel {
 }
 
 export interface Output extends Channel {}
+
+export function isComponentChannel(channel: Channel): boolean {
+  return Boolean(channel.id) && !channel.id.startsWith("@");
+}
+
+export function isHostChannel(channel: Channel): boolean {
+  return channel.id === "@app";
+}
+
+export function isContainerChannel(channel: Channel): boolean {
+  return channel.id === "@container";
+}